♻️(frontend) extract getTrackKey to shared layout utils
Move track key helper out of pip so layouts can reuse it.
